feat: 폴더 소유자 이미지 없을 시 기본 이미지 적용
