+++ src/App.css
@@ -3,7 +3,7 @@
 .App header {
   position: sticky;
   top: 0;
-  z-index: 1;
+  z-index: 10;
 }
 
 .App .content_container {
