+++ src/components/main/Cards.js
@@ -1,4 +1,5 @@
 import "./Cards.css";
+import noCardImg from "../../assets/bg_noImg.png";
 
 function Cards({ cardData }) {
   // 시간 차 구하기
@@ -25,7 +26,19 @@ function Cards({ cardData }) {
         {cardData.map((elem) => (
           <a href={elem.url} target="_blank" className="card" key={elem.id}>
             <div className="card_image_container">
-              <img className="card_image" src={elem.imageSource} />
+              {elem.imageSource ? (
+                <img
+                  className="card_image"
+                  src={elem.imageSource}
+                  alt="card image"
+                />
+              ) : (
+                <img
+                  className="card_image"
+                  src={noCardImg}
+                  alt="no card image"
+                />
+              )}
             </div>
             <div className="card_information">
               <div className="createdAt">{ElapsedTime(elem)}</div>
