+++ src/components/main/FolderInformation.css
@@ -4,7 +4,7 @@
   box-sizing: border-box;
 }
 
-.UserInformation {
+.FolderInformation {
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -13,12 +13,12 @@
   background: var(--Gray5);
 }
 
-.UserInformation .user_image {
+.FolderInformation .user_image {
   width: 60px;
   height: 60px;
 }
 
-.UserInformation .folder_owner {
+.FolderInformation .folder_owner {
   margin-top: 12px;
   margin-bottom: 20px;
   font-family: Pretendard;
@@ -27,7 +27,7 @@
   line-height: 24px;
 }
 
-.UserInformation .folder_name {
+.FolderInformation .folder_name {
   font-family: Pretendard;
   font-size: 40px;
   font-weight: 600;
